Added `FilesUtility::resetTmpPath()' and used at `Server' initialization.  Some compiler warnings silented.
git-svn-id: svn://svn.savannah.gnu.org/myserver/trunk@2893 c4c922ea-a70d-0410-937d-ec1daf91936c

+++ myserver/src/http_handler/cgi/cgi.cpp
@@ -360,7 +360,7 @@ int Cgi::sendData (HttpThreadContext* td, Pipe &stdOutFile, FiltersChain& chain,
       int aliveProcess = 0;
       u_long ticks = getTicks() - procStartTime;
 
-      if (ticks >= cgiTimeout || 
+      if (ticks >= (u_long)cgiTimeout || 
           stdOutFile.waitForData ((cgiTimeout - ticks) / 1000, (cgiTimeout - ticks) % 1000) == 0)
       {
         ostringstream msg;
